refactor(backend): split archivos monolíticos + quick wins de clean code
Archivos divididos en módulos:
- billing/models.py (952 líneas) → billing/models/ (6 archivos)
- core/admin.py (1444 líneas) → core/admin/ (6 archivos)
- websites/services/ai_service.py (1028 líneas) → 5 archivos

Quick wins:
- Centralizar _get_client_ip() en core/utils (eliminar 2 duplicados)
- Crear safe_delay() wrapper para Celery (reemplazar 5 try/delay/except)
- Agregar logging a 8 except Exception: pass silenciosos

Backward compatibility garantizada vía re-exports en __init__.py.
No se modificó lógica de negocio ni migraciones.

Closes #243
