+++ backend/billing/models/__init__.py
@@ -1,0 +1,24 @@
+# backend/billing/models/__init__.py
+"""
+Sistema de Billing Modular para NERBIS.
+
+Re-exporta todos los modelos para backward compatibility.
+Cualquier ``from billing.models import X`` existente sigue funcionando.
+"""
+
+from .modules import Module
+from .plans import Plan
+from .pricing import PricingConfig
+from .subscriptions import Subscription, SubscriptionModule
+from .usage import Invoice, InvoiceLineItem, UsageRecord
+
+__all__ = [
+    "Invoice",
+    "InvoiceLineItem",
+    "Module",
+    "Plan",
+    "PricingConfig",
+    "Subscription",
+    "SubscriptionModule",
+    "UsageRecord",
+]
